Classes

+++ Classes/src/Challenges/SimpleCalculator/VipCustomer.java
@@ -2,17 +2,17 @@ package Challenges.SimpleCalculator;
 
 public class VipCustomer {
     private String name;
-    private int creditLimit;
+    private double creditLimit;
     private String email;
 
 
     public VipCustomer(){
         this("Default name",0,"Default@email.com");
     }
-    public VipCustomer(String name,int creditLimit){
+    public VipCustomer(String name,double creditLimit){
         this(name,creditLimit,"default@gmail.com");
     }
-    public VipCustomer(String name,int creditLimit,String email){
+    public VipCustomer(String name,double creditLimit,String email){
         this.name=name;
         this.creditLimit=creditLimit;
         this.email=email;
@@ -21,7 +21,7 @@ public class VipCustomer {
         return name;
     }
 
-    public int getCreditLimit() {
+    public double getCreditLimit() {
         return creditLimit;
     }
